#25: Set the pieces
The end goal is to decouple RootTabBarController from controlling itself. This will help us in the future if we want to change behavior (e.g. change behavior if the user is logged in or not) in a very simple and effective manner.

For now, move the code to where it should be and we'll go from there.