#26: Different instances
As you can see, at the crash site (`RootTabBarController.swift:63) our instance is the same as the instance at `installRootTabBar` (`Navigator.swift:66) _but_ it's different when calling `rootTabBarController()` (`Navigator.swift:139).

It's interesting. I think we're instantiating the rootTabBarController from a storyboard and then passing that VC to the view hierarchy, but we are never setting its delegate. The new object in `rootTabBarController` is part of `Navigator` and is never in the view hierarchy...

We can test it by looking for the storyboard and setting its action delegate somewhere.